Move findUser and retrieveUsers into user.ts
